Combined Feature Engi & Model Building
We have made several modifications to the original Feature Engeneering file of Gokul. After putting the new variables in the model, severals are automatically excluded by the model, because they are the linear combinaition of the existant features, such as "HasXX" is actually a dummp variable that R will aotomatically create in the lm model. So we removed it and generated the new Test and Train csv, named as Newest.csv